Messaging: redesign user callback API
Rather than having the user register individual delegate callback
functions, let the user, if they choose to, provide an object
which will be called when various notifications or events need to
be published to the user.

There is a default user implementation, so the user need not override
any or all of the actual callbacks.