MessagingTests: fix race condition
AwaitMessage would sample the number of received messages
and then wait for the number of received messages to change.
However, the sampling would occur _after_ the process to
receive a message had started, so there was a race between
receiving the message and sampling the number of messages
received.

Replace AwaitMessage with AwaitMessages and have it take in
the number of received messages to await.